+++ generic-extras/src/test/scala/io/circe/generic/extras/EnumerationSemiautoDerivedSuite.scala
